Fix type of $timeout parameter of td_json_client_receive
Php::Type::Numeric is an integer, not a float.
See:
- https://github.com/yaroslavche/phptdlib/blob/0.0.10/tdlib.cpp#L159
- https://github.com/CopernicaMarketingSoftware/PHP-CPP/blob/v2.2.0/include/type.h#L25

+++ stubs/td_json_client.php
@@ -16,6 +16,6 @@ function td_json_client_send(TDLib\BaseJsonClient $client, string $query)
 {
 }
 
-function td_json_client_receive(TDLib\BaseJsonClient $client, float $timeout)
+function td_json_client_receive(TDLib\BaseJsonClient $client, int $timeout)
 {
 }
